kibana_fetch_logs: Adapt according to current indexed data
Prior to this, the args/kwargs/exception were directly retrieved from
the swh_logging_{args|kwargs|exception} field. This is no longer the
case.

What possibly changed this was the migration between banco to the
elasticsearch cluster... When we were using banco, logstash might have
transformed the data. I no longer think we use a logstash to push the
logs?