elastic-worker: Deal with amqp credentials as secrets
This currently read the secrets and installs them properly in our config.yml the loader
uses to ingest origins.

It's currently convoluted but that's the simpler way for now. Short of changing the swh
tids and bits (loader, lister, ...) to actually allow using environment variables
instead of configuration here (sure but not right now).

Related to T4144